+++ app/Discord/FlowMeasure/Content/FlowMeasureRecipientsInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Discord\FlowMeasure\Content;
+
+interface FlowMeasureRecipientsInterface
+{
+    public function toString(): string;
+}
