+++ app/Discord/FlowMeasure/Message/FlowMeasureNotifiedMessage.php
@@ -2,7 +2,7 @@
 
 namespace App\Discord\FlowMeasure\Message;
 
-use App\Discord\FlowMeasure\Content\InterestedParties;
+use App\Discord\FlowMeasure\Content\EcfmpInterestedParties;
 use App\Discord\FlowMeasure\Description\EventName;
 use App\Discord\FlowMeasure\Field\ArrivalAirports;
 use App\Discord\FlowMeasure\Field\DepartureAirports;
@@ -34,7 +34,7 @@ class FlowMeasureNotifiedMessage implements MessageInterface
 
     public function content(): string
     {
-        return InterestedParties::interestedPartiesString($this->measure);
+        return EcfmpInterestedParties::interestedPartiesString($this->measure);
     }
 
     public function embeds(): EmbedCollection
